New movements mechanism
New sets of functions and constants to make the robot move according to time.
This is not very accurate, but is a fix for broken encoders readings.

To change the values, play with SPEED (line 60) - this should be enough to calibrate it. The current base value set at 40cm/s is an estimate.